feat(general): add support for running only a part of the publish run.
The following object property has been added to publishrc.json to achieve this functionality:

    republish: { }

When this property is defined, and contains at least one of it's own properties, the publish will process only the tasks specified within the republish object.  History/changelog files will not be opened for editing and the version used will be the current one.  For example, to make a MantisBT release only, the republish property can be set as follows:

    republish:
    {
        mantisbtRelease: "Y"
    }