feat(General): Add support for version replacement using custom tag.
This is acheived via the following new property in .publishrc.json:

    versionReplaceTags

This new property can be specified as a string, or an array of strings. that can be used for version updating in custom version files defined in the "versionFiles" property of .publishrc.json.

For example:

Aassume the new version number is 1.1.0 and the old version number is 1.0.5, and the following is defined in .publishrc.json:

     "versionReplaceTags": "PROJECT_VERSION = "

This will cause any occurences of "DATABASE_VERSION = 1.0.5" in the files specified in "versionFiles" will be replaced with the string "PROJECT_VERSION = 1.1.0" .