refactor: place 'build' commit comments at the end of the generated changelog/history file.
fix: if a history file already exists when aking a first release, using the 'interactive' version input mechanism causes the history file to not be written with the appropriate version header information.

fix: detect a first release and do not increment the version that was read in from the version file (package.json/history.txt/.publishrc.json).