feat: Add new 'task mode' command line switch:
    -tvi ,  --task-version-info

Using this option instructs the process to output both the current version of the last released build, and the proposed next version # to stdout.  Outputs in the format 'current|next'.

feat: Add new 'task mode' command line switch:

    -tcei,  --task-ci-env-info

Using this option instructs the process to output CI related buildinformation to stdout.  Outputs in the format 'current|next|changelog'.

feat: Add new command line switch option:

    --version-force-current ,  --version-force

Using this option instructs the process to use the provided version number for the 'current version'.

feat: Add new command line switch option:

    --version-force-next

Using this option instructs the process to use the provided version number for the 'next version'.  Note that a version number will not be calculated, and a VCS request will be skipped.